cram/record: Change ID to a u64

+++ noodles-cram/src/data_container/slice.rs
@@ -98,8 +98,8 @@ impl Slice {
         let record_count = self.header().record_count();
         let mut records = Vec::with_capacity(record_count);
 
-        let start_id = self.header().record_counter();
-        let end_id = start_id + (record_count as i64);
+        let start_id = self.header().record_counter() as u64;
+        let end_id = start_id + (record_count as u64);
 
         for id in start_id..end_id {
             let mut record = record_reader.read_record()?;
